added problems to lobbies
