--- src/routes/lobby/[id]/+page.svelte
@@ -1,5 +1,0 @@
-<div>
-    <div class="text-center text-2xl font-bold">
-        Loading...
-    </div>
-</div>
